make external auth more complete
Now has built-in route for checking login request,
approving login request and waiting for approval.

The "wait" one waits for approval and polls the "check" route
to see if it should finally redirect to `loginDest`. This is all managed
by the plugin, allowing me
to remove most of the code for it.

The template remains; I'm not sure if I can ship it with the plugin
separately.